refactor: modify logic of AlertModal open
